kmkfrontpage: more robust searching for "date"
The gallery should write out metadata about itself on creation, but, well, it
doesn't.  We're scanning html text for the last file, instead of looking for
exif because the images can/are stripped, but after layout changes, the simple
dump to html and regexp was failing, so move to a more structure digging
attempt to find the right fields.