pyexiftool0.5+ and iptc with utf8
just slam all this into one commit like an amateur.

We discovered that some of our old images had iptc with utf8 encoding,
which was _probably_ a bug in a 20 year old version of exiftool when we
were writing the files, but no matter, add the flags that let us
selectively handle that conversion, and thank the authors that it's just
a config option.  We now have the tools to more properly recreate any
old gallery that we find errors on.